@@ -14,8 +14,6 @@ classifier =
     License :: OSI Approved :: Apache Software License
     Operating System :: OS Independent
     Programming Language :: Python
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
@@ -27,7 +25,7 @@ project_urls =
     Documentation = https://stestr.readthedocs.io
     Source Code = https://github.com/mtreinish/stestr
     Bug Tracker = https://github.com/mtreinish/stestr/issues
-requires-python = >=2.7
+requires-python = >=3.5
 
 [files]
 packages =
@@ -48,9 +46,6 @@ stestr.cm =
 [extras]
 sql =
     subunit2sql>=1.8.0
-
-[bdist_wheel]
-universal=1
 
 [build_sphinx]
 source-dir = doc/source
