@@ -10,10 +10,9 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+import io
 import re
-
-import mock
-import six
+from unittest import mock
 
 from stestr import selection
 from stestr.tests import base
@@ -41,12 +40,12 @@ class TestSelection(base.TestCase):
 
 class TestBlackReader(base.TestCase):
     def test_black_reader(self):
-        blacklist_file = six.StringIO()
+        blacklist_file = io.StringIO()
         for i in range(4):
             blacklist_file.write('fake_regex_%s\n' % i)
             blacklist_file.write('fake_regex_with_note_%s # note\n' % i)
         blacklist_file.seek(0)
-        with mock.patch('six.moves.builtins.open',
+        with mock.patch('builtins.open',
                         return_value=blacklist_file):
             result = selection.black_reader('fake_path')
         self.assertEqual(2 * 4, len(result))
@@ -60,10 +59,10 @@ class TestBlackReader(base.TestCase):
         self.assertEqual(note_cnt, 4)
 
     def test_invalid_regex(self):
-        blacklist_file = six.StringIO()
+        blacklist_file = io.StringIO()
         blacklist_file.write("fake_regex_with_bad_part[The-BAD-part]")
         blacklist_file.seek(0)
-        with mock.patch('six.moves.builtins.open',
+        with mock.patch('builtins.open',
                         return_value=blacklist_file):
             with mock.patch('sys.exit') as mock_exit:
                 selection.black_reader('fake_path')
@@ -108,13 +107,13 @@ class TestConstructList(base.TestCase):
             result = selection.construct_list(test_lists,
                                               whitelist_file='file')
         self.assertEqual(set(result),
-                         set(('fake_test1[tg]', 'fake_test2[tg]')))
+                         {'fake_test1[tg]', 'fake_test2[tg]'})
 
     def test_whitelist_invalid_regex(self):
-        whitelist_file = six.StringIO()
+        whitelist_file = io.StringIO()
         whitelist_file.write("fake_regex_with_bad_part[The-BAD-part]")
         whitelist_file.seek(0)
-        with mock.patch('six.moves.builtins.open',
+        with mock.patch('builtins.open',
                         return_value=whitelist_file):
             with mock.patch('sys.exit') as mock_exit:
                 selection._get_regex_from_whitelist_file('fake_path')
@@ -133,7 +132,7 @@ class TestConstructList(base.TestCase):
                 result = selection.construct_list(test_lists, 'black_file',
                                                   'white_file', ['foo'])
         self.assertEqual(set(result),
-                         set(('fake_test1[tg]', 'fake_test3[tg,foo]')))
+                         {'fake_test1[tg]', 'fake_test3[tg,foo]'})
 
     def test_overlapping_black_regex(self):
 
