@@ -10,9 +10,9 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+import io
 import sys
 
-import six
 import subunit
 import testtools
 
@@ -62,7 +62,7 @@ def output_table(table, output=sys.stdout):
         outputs.append('  ')
     for row in contents[1:]:
         show_row(row)
-    output.write(six.text_type('').join(outputs))
+    output.write(''.join(outputs))
 
 
 def output_tests(tests, output=sys.stdout):
@@ -75,8 +75,8 @@ def output_tests(tests, output=sys.stdout):
 
     for test in tests:
         id_str = test.id()
-        output.write(six.text_type(id_str))
-        output.write(six.text_type('\n'))
+        output.write(str(id_str))
+        output.write('\n')
 
 
 def make_result(get_id, output=sys.stdout):
@@ -109,16 +109,16 @@ def output_summary(successful, tests, tests_delta, time, time_delta, values,
     summary = []
     a = summary.append
     if tests:
-        a("Ran %s" % (tests,))
+        a("Ran {}".format(tests))
         if tests_delta:
             a(" (%+d)" % (tests_delta,))
         a(" tests")
     if time:
         if not summary:
             a("Ran tests")
-        a(" in %0.3fs" % (time,))
+        a(" in {:0.3f}s".format(time))
         if time_delta:
-            a(" (%+0.3fs)" % (time_delta,))
+            a(" ({:+0.3f}s)".format(time_delta))
     if summary:
         a("\n")
     if successful:
@@ -129,13 +129,13 @@ def output_summary(successful, tests, tests_delta, time, time_delta, values,
         a(' (')
         values_strings = []
         for name, value, delta in values:
-            value_str = '%s=%s' % (name, value)
+            value_str = '{}={}'.format(name, value)
             if delta:
                 value_str += ' (%+d)' % (delta,)
             values_strings.append(value_str)
         a(', '.join(values_strings))
         a(')')
-    output.write(six.text_type(''.join(summary)) + six.text_type('\n'))
+    output.write(''.join(summary) + '\n')
 
 
 def output_stream(stream, output=sys.stdout):
@@ -169,20 +169,20 @@ class ReturnCodeToSubunit(object):
         self.proc = process
         self.done = False
         self.source = self.proc.stdout
-        self.lastoutput = six.binary_type(('\n').encode('utf8')[0])
+        self.lastoutput = bytes((b'\n')[0])
 
     def _append_return_code_as_test(self):
         if self.done is True:
             return
-        self.source = six.BytesIO()
+        self.source = io.BytesIO()
         returncode = self.proc.wait()
         if returncode != 0:
-            if self.lastoutput != six.binary_type(('\n').encode('utf8')[0]):
+            if self.lastoutput != bytes((b'\n')[0]):
                 # Subunit V1 is line orientated, it has to start on a fresh
                 # line. V2 needs to start on any fresh utf8 character border
                 # - which is not guaranteed in an arbitrary stream endpoint, so
                 # injecting a \n gives us such a guarantee.
-                self.source.write(six.binary_type('\n'))
+                self.source.write(bytes('\n'))
             stream = subunit.StreamResultToBytes(self.source)
             stream.status(test_id='process-returncode', test_status='fail',
                           file_name='traceback',
@@ -194,7 +194,7 @@ class ReturnCodeToSubunit(object):
 
     def read(self, count=-1):
         if count == 0:
-            return six.text_type('')
+            return ''
         result = self.source.read(count)
         if result:
             self.lastoutput = result[-1]
