@@ -10,10 +10,10 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import mock
 import os
 import shutil
 import tempfile
+from unittest import mock
 
 from stestr.repository import util
 from stestr.tests import base
