@@ -12,9 +12,8 @@
 
 import io
 import operator
+from unittest import mock
 
-import mock
-import six
 import subunit
 import testtools
 
@@ -80,7 +79,7 @@ class FakeNoFailing(FakeTestRun):
 
     def __init__(self, failure=True):
         # Generate a subunit stream
-        stream_buf = io.BytesIO(six.binary_type(''.encode('utf-8')))
+        stream_buf = io.BytesIO(bytes(b''))
         self._content = stream_buf.getvalue()
         self.id = None
 
