@@ -75,8 +75,8 @@ class UserConfig(object):
         try:
             self.schema(self.config)
         except vp.MultipleInvalid as e:
-            msg = 'Provided user config file %s is invalid because:\n%s' % (
-                path, str(e))
+            msg = ('Provided user config file {} is invalid '
+                   'because:\n{}'.format(path, str(e)))
             sys.exit(msg)
 
     @property
